Struct rplex::Problem

pub struct Problem<'a> { /* fields omitted */ }

A CPLEX Problem.

Implementations

impl<'a> Problem<'a>

pub fn new<S>(env: &'a Env, name: S) -> Result<Self, String> where
S: Into<String>,

pub fn add_variable(&mut self, var: Variable) -> Result<usize, String>

Add a variable to the problem. The Variable is moved into the problem. At this time, it is not possible to get a reference to it back.

The column index for the Variable is returned.

pub fn add_constraint(&mut self, con: Constraint) -> Result<usize, String>

Add a constraint to the problem.

The row index for the constraint is returned.

pub fn add_lazy_constraint(&mut self, con: Constraint) -> Result<usize, String>

Adds a lazy constraint to the problem.

Returns the index of the constraint. Unclear if this has any value whatsoever.

pub fn set_objective(
&mut self,
ty: ObjectiveType,
con: Constraint
) -> Result<(), String>

Set the objective coefficients. A Constraint object is used because it encodes a weighted sum, although it is semantically incorrect. The right-hand-side and kind of (in)equality of the Constraint are ignored.

pub fn set_objective_type(&mut self, ty: ObjectiveType) -> Result<(), String>

Change the objective type. Default: ObjectiveType::Minimize.

It is recommended to use this in conjunction with objective coefficients set by the var! macro rather than using set_objective.

pub fn write<S>(&self, name: S) -> Result<(), String> where
S: Into<String>,

Write the problem to a file named name. At this time, it is not possible to use a Write object instead, as this calls C code directly.

pub fn add_initial_soln(
&mut self,
vars: &[usize],
values: &[f64]
) -> Result<(), String>

Add an initial solution to the problem.

vars is an array of indices (i.e. the result of prob.add_variable) and values are their values.

pub fn solve_as(&mut self, pt: ProblemType) -> Result<Solution, String>

Solve the Problem, returning a Solution object with the result.

pub fn solve(&mut self) -> Result<Solution, String>

Solve the problem as a Mixed Integer Program